Interface to assist with unmarshalling complex NDR structures
Marshalling of any data type via NDR can be decomposed into three stages:
 * Preamble
 * Entity
 * Deferrals

When each of these are called is dependant on the parent data type.
The steps taken during each step is up to the implementation of the data type.

A convenience method for marshalling and unmarshalling DataType instances is now provided in PacketOutput and PacketInput respectively.