Introducing new alignment strategy
Previously, all `DataType`'s needed to report their alignment sizes via `Alignment getAlignment()`. Using this, it was the caller's responsibility to align the stream before writing the object.

However, this only works for `marshalEntity`. Separate alignments for both the preamble and deferrals must be considered.

To accomodate this, instead of introducing two more getters, the `DataType`'s are now responsible for their own alignment. This comes at a very small CPU cost when the stream is already known to be aligned, and alignment was not required, but greatly simplifies the reuse of structs within the framework.

- All existing `DataType` implementations now align themselves to the appropriate alignment in all three stages: preamble, entity, and deferrals
- Unit tests have been added to ensure that alignment takes place where expected
- Updating README.md with new alignment strategy. Adding additional examples